requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bert-score==0.3.13
bitsandbytes==0.41.1
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.6
dbus-python==1.2.18
dill==0.3.7
distro==1.7.0
evaluate==0.4.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.18.0
idna==3.4
importlib-resources==6.1.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.1.0
lightning-utilities==0.9.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.2
peft==0.6.0
Pillow==10.1.0
psutil==5.9.6
pyarrow==14.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
safetensors==0.4.0
scipy==1.11.3
SecretStorage==3.3.1
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
transformers==4.33.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wadllib==1.3.6
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
